requirements.txt
passlib
python-jose
sqlalchemy
mysql-connector-python==8.1.0
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.3.0
PyJWT==2.8.0
